+++ x-pack/plugins/lens/public/datatable_visualization/datatable_visualization.ts
@@ -5,6 +5,6 @@
  * 2.0.
  */
 
-export { datatableColumn, getDatatable } from '../../common/expressions';
+export { datatableColumn, getDatatableTransformer, datatable } from '../../common/expressions';
 export * from './expression';
 export * from './visualization';
